Move timeline to separate component
